tasker needs styling

+++ js/Login_Reg.js
@@ -1,30 +1,35 @@
 
 const signUp = e => {
+  
     let email = document.getElementById('emailInput').value,
-        pwd = document.getElementById('floatingPwd').value,
-        confirmaton = document.getElementById('pwdConfirm').value;
-        
+    pwd = document.getElementById('floatingPwd').value,
+    confirmaton = document.getElementById('pwdConfirm').value;
 
         // campare password and confirmation paswword
         if(pwd!=confirmaton){
             alert("does not match");
             return false
         }
+        if(pwd==""){
+            alert("Please fill in the missing fields");
+            return true
+        }
         
-   
+        //create the key [] for the localstorage
         let formData = JSON.parse(localStorage.getItem('formData')) || [];
 
         let exist = formData.length && 
         JSON.parse(localStorage.getItem('formData')).some(data => 
         data.email == email && 
         data.pwd == pwd
+
         );
     
      if(!exist){
-         formData.push({ email, pwd, confirmaton });
+         formData.unshift({ email, pwd, confirmaton });
          localStorage.setItem('formData', JSON.stringify(formData));
          document.querySelector('.theForm').reset(); // resetting the form
-         //document.getElementById('').focus();
+         document.getElementById('').focus();
          alert("Account Created.\n\n Please continue.");
          location.href = "/index.html";
      }
